File structure, imports, to be able to use run.py, and to work with broken slack
