+++ run.py
@@ -1,0 +1,4 @@
+from nom_track_app.app import app
+
+if __name__ == '__main__':
+    app.run()
